finish rewrite, everything but docset toc
This should complete the rewrite from posh-to-dash and into a more correct toc downloader to support further msdocs.
Code to support static downloading and manipulation, exists but now we favor in-memory and writing to tar directly.
these really need to be kept for larger document storage tasks and maybe should be handled at docset config level.

currently in testing of sqlite, icons, plist, and tar